backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sf-tracking-dashboard"
version = "0.1.0"
description = "SF Express local tracking dashboard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1",
  "apscheduler>=3.10.4",
  "cryptography>=44.0.0",
  "fastapi>=0.115.11",
  "httpx>=0.28.1",
  "openpyxl>=3.1.5",
  "pandas>=2.2.3",
  "pydantic-settings>=2.7.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.38",
  "truststore>=0.10.1",
  "uvicorn[standard]>=0.34.0",
  "xlrd>=2.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-cov>=6.0.0",
  "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
